requirements.txt
sklearn
tensorflow-gpu==2.1
transformers==2.11.0
